# server/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Fantasy FRC API
  description: |
    Machine-to-machine API for automating Fantasy FRC drafts.

    Authentication uses a short-lived JWT obtained by exchanging a client id
    and secret via `POST /api/v1/auth/token`. Include the JWT in the
    `Authorization: Bearer <token>` header on all other requests.
  version: 1.0.0
  contact:
    name: Fantasy FRC
servers:
  - url: /api/v1
    description: API v1
security: []
paths:
...
  /drafts/{id}:
    parameters:
      - $ref: '#/components/parameters/DraftId'

    get:
      operationId: getDraft
      summary: Get a single draft
      description: Requires membership in the draft.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Draft details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      operationId: updateDraft
      summary: Update draft metadata
      description: Only the draft owner may update a draft.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDraftRequest'
      responses:
        '200':
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /drafts/{id}/picks:
    parameters:
      - $ref: '#/components/parameters/DraftId'
    get:
      operationId: listPicks
      summary: List all picks in a draft
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of picks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pick'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: makePick
      summary: Make a pick
      description: |
        Submit a team number as the current picking player. The team number
        will be prefixed with `frc`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MakePickRequest'
      responses:
        '200':
          description: Pick made
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pick'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /drafts/{id}/score:
    parameters:
      - $ref: '#/components/parameters/DraftId'
    get:
      operationId: getDraftScore
      summary: Get the score breakdown for a draft
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Draft score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftScore'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/token
  parameters:
    DraftId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Draft ID
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error description
...
    UserSummary:
      type: object
      required:
        - uuid
        - username
      properties:
        uuid:
          type: string
          format: uuid
        username:
          type: string
...
    DraftState:
      type: string
      enum:
        - Filling
        - Waiting to Start
        - Picking
        - Teams Playing
        - Complete
...
    UpdateDraftRequest:
      type: object
      properties:
        display_name:
          type: string
        description:
          type: string
        interval:
          type: integer
          description: Number of seconds per pick
    Draft:
      type: object
      required:
        - id
        - display_name
        - description
        - interval
        - owner
        - status
        - players
      properties:
        id:
          type: integer
        display_name:
          type: string
        description:
          type: string
        interval:
          type: integer
        owner:
          $ref: '#/components/schemas/UserSummary'
        status:
          $ref: '#/components/schemas/DraftState'
        players:
          type: array
          items:
            $ref: '#/components/schemas/DraftPlayer'
        next_pick:
          $ref: '#/components/schemas/DraftPlayer'
    DraftPlayer:
      type: object
      required:
        - id
        - user
        - player_order
        - pending
        - score
      properties:
        id:
          type: integer
        user:
          $ref: '#/components/schemas/UserSummary'
        player_order:
          type: integer
        pending:
          type: boolean
        score:
          type: integer
        picks:
          type: array
          items:
            $ref: '#/components/schemas/Pick'
    Pick:
      type: object
      required:
        - id
        - player_id
        - available_time
        - expiration_time
        - skipped
        - score
      properties:
        id:
          type: integer
        player_id:
          type: integer
        team_key:
          type: string
          description: TBA team key, e.g. frc254
        pick_time:
          type: string
          format: date-time
        available_time:
          type: string
          format: date-time
        expiration_time:
          type: string
          format: date-time
        skipped:
          type: boolean
        score:
          type: integer
    MakePickRequest:
      type: object
      required:
        - team_number
      properties:
        team_number:
          type: integer
...
    DraftScore:
      type: object
      required:
        - draft_id
        - status
        - players
      properties:
        draft_id:
          type: integer
        status:
          $ref: '#/components/schemas/DraftState'
        players:
          type: array
          items:
            $ref: '#/components/schemas/DraftPlayer'
...